requirements.txt
anyio==4.12.1
boto3==1.42.47
botocore==1.42.47
certifi==2026.1.4
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
urllib3==2.0.7
